@@ -38,10 +38,16 @@
     }
   },
   "dependencies": {
-    "@map-colonies/error-express-handler": "^1.2.0",
-    "@map-colonies/mc-logger": "^1.2.1",
-    "@map-colonies/mc-probe": "^1.0.0",
+    "@godaddy/terminus": "^4.7.1",
+    "@map-colonies/error-express-handler": "^2.0.0",
+    "@map-colonies/express-access-log-middleware": "0.0.3",
+    "@map-colonies/js-logger": "0.0.3",
     "@map-colonies/openapi-express-viewer": "^2.0.1",
+    "@map-colonies/telemetry": "^1.1.0",
+    "@opentelemetry/api": "1.0.0-rc.0",
+    "@opentelemetry/api-metrics": "0.19.0",
+    "@opentelemetry/instrumentation-express": "0.16.0",
+    "@opentelemetry/instrumentation-http": "0.19.0",
     "compression": "^1.7.4",
     "config": "^3.3.6",
     "express": "^4.17.1",
@@ -82,6 +88,6 @@
     "standard-version": "^9.0.0",
     "supertest": "^6.0.1",
     "ts-jest": "^26.3.0",
-    "typescript": "^3.8.3"
+    "typescript": "^4.2.4"
   }
 }
