@@ -1,9 +1,9 @@
 import { RequestHandler } from 'express';
 import httpStatus from 'http-status-codes';
 import { injectable, inject } from 'tsyringe';
+import { Logger } from '@map-colonies/js-logger';
 import { HttpError } from '@map-colonies/error-express-handler';
 import { Services } from '../../common/constants';
-import { ILogger } from '../../common/interfaces';
 
 import { MergedModel, MergedIdMapping, MergeManager } from '../models/mergerManager';
 
@@ -11,7 +11,7 @@ type MergeHandler = RequestHandler<undefined, MergedModel[], MergedIdMapping>;
 
 @injectable()
 export class MergeController {
-  public constructor(@inject(Services.LOGGER) private readonly logger: ILogger, @inject(MergeManager) private readonly manager: MergeManager) {}
+  public constructor(@inject(Services.LOGGER) private readonly logger: Logger, @inject(MergeManager) private readonly manager: MergeManager) {}
   public merge: MergeHandler = (req, res) => {
     let result;
     try {
