@@ -1,24 +1,21 @@
 import express from 'express';
 import bodyParser from 'body-parser';
 import compression from 'compression';
+import { Logger } from '@map-colonies/js-logger';
+import httpLogger from '@map-colonies/express-access-log-middleware';
 import { getErrorHandlerMiddleware } from '@map-colonies/error-express-handler';
 import { OpenapiViewerRouter, OpenapiRouterConfig } from '@map-colonies/openapi-express-viewer';
 import { middleware as OpenApiMiddleware } from 'express-openapi-validator';
 import { container, inject, injectable } from 'tsyringe';
-import { RequestLogger } from './common/middlewares/RequestLogger';
 import { Services } from './common/constants';
-import { IConfig, ILogger } from './common/interfaces';
+import { IConfig } from './common/interfaces';
 import { mergeRouterFactory } from './merger/routes/mergerRouter';
 
 @injectable()
 export class ServerBuilder {
-  private readonly serverInstance = express();
+  private readonly serverInstance: express.Application;
 
-  public constructor(
-    @inject(Services.CONFIG) private readonly config: IConfig,
-    private readonly requestLogger: RequestLogger,
-    @inject(Services.LOGGER) private readonly logger: ILogger
-  ) {
+  public constructor(@inject(Services.CONFIG) private readonly config: IConfig, @inject(Services.LOGGER) private readonly logger: Logger) {
     this.serverInstance = express();
   }
 
@@ -45,16 +42,16 @@ export class ServerBuilder {
     if (this.config.get<boolean>('server.response.compression.enabled')) {
       this.serverInstance.use(compression(this.config.get<compression.CompressionFilter>('server.response.compression.options')));
     }
+    this.serverInstance.use(httpLogger({ logger: this.logger }));
+
     this.serverInstance.use(express.json(this.config.get<bodyParser.Options>('server.request.payload')));
 
     const ignorePathRegex = new RegExp(`^${this.config.get<string>('openapiConfig.basePath')}/.*`, 'i');
     const apiSpecPath = this.config.get<string>('openapiConfig.filePath');
     this.serverInstance.use(OpenApiMiddleware({ apiSpec: apiSpecPath, validateRequests: true, ignorePaths: ignorePathRegex }));
-
-    this.serverInstance.use(this.requestLogger.getLoggerMiddleware());
   }
 
   private registerPostRoutesMiddleware(): void {
-    this.serverInstance.use(getErrorHandlerMiddleware((message) => this.logger.log('error', message)));
+    this.serverInstance.use(getErrorHandlerMiddleware());
   }
 }
