@@ -1,18 +1,28 @@
-import { readFileSync } from 'fs';
 import { container } from 'tsyringe';
 import config from 'config';
-import { Probe } from '@map-colonies/mc-probe';
-import { MCLogger, ILoggerConfig, IServiceConfig } from '@map-colonies/mc-logger';
+import { logMethod, Tracing, Metrics } from '@map-colonies/telemetry';
+import jsLogger, { LoggerOptions } from '@map-colonies/js-logger';
 import { Services } from './common/constants';
 
-function registerExternalValues(): void {
-  const loggerConfig = config.get<ILoggerConfig>('logger');
-  const packageContent = readFileSync('./package.json', 'utf8');
-  const service = JSON.parse(packageContent) as IServiceConfig;
-  const logger = new MCLogger(loggerConfig, service);
+function registerExternalValues(tracing: Tracing): void {
+  const loggerConfig = config.get<LoggerOptions>('logger');
+  // @ts-expect-error the signature is wrong
+  const logger = jsLogger({ ...loggerConfig, prettyPrint: false, hooks: { logMethod } });
   container.register(Services.CONFIG, { useValue: config });
   container.register(Services.LOGGER, { useValue: logger });
-  container.register<Probe>(Probe, { useFactory: (container) => new Probe(container.resolve(Services.LOGGER), {}) });
+
+  const metrics = new Metrics('id-merger');
+  const meter = metrics.start();
+  container.register(Services.METER, { useValue: meter });
+
+  const tracer = tracing.start();
+  container.register(Services.TRACER, { useValue: tracer });
+
+  container.register('onSignal', {
+    useValue: async (): Promise<void> => {
+      await Promise.all([tracing.stop(), metrics.stop()]);
+    },
+  });
 }
 
 export { registerExternalValues };
