@@ -1,10 +1,21 @@
+/* eslint-disable import/first */
 // this import must be called before the first import of tsyring
 import 'reflect-metadata';
-import { Probe } from '@map-colonies/mc-probe';
+import { Tracing } from '@map-colonies/telemetry';
+import { createTerminus } from '@godaddy/terminus';
+import { HttpInstrumentation } from '@opentelemetry/instrumentation-http';
+import { ExpressInstrumentation } from '@opentelemetry/instrumentation-express';
+import { Logger } from '@map-colonies/js-logger';
 import { container } from 'tsyringe';
 import { get } from 'config';
+import { DEFAULT_SERVER_PORT, IGNORED_INCOMING_TRACE_ROUTES, IGNORED_OUTGOING_TRACE_ROUTES, Services } from './common/constants';
+
+const tracing = new Tracing('id-merger', [
+  new HttpInstrumentation({ ignoreOutgoingUrls: IGNORED_OUTGOING_TRACE_ROUTES, ignoreIncomingPaths: IGNORED_INCOMING_TRACE_ROUTES }),
+  new ExpressInstrumentation(),
+]);
+
 import { getApp } from './app';
-import { DEFAULT_SERVER_PORT } from './common/constants';
 
 interface IServerConfig {
   port: string;
@@ -12,8 +23,12 @@ interface IServerConfig {
 
 const serverConfig = get<IServerConfig>('server');
 const port: number = parseInt(serverConfig.port) || DEFAULT_SERVER_PORT;
-const app = getApp();
-const probe = container.resolve<Probe>(Probe);
-void probe.start(app, port).then(() => {
-  probe.readyFlag = true;
+
+const app = getApp(tracing);
+
+const logger = container.resolve<Logger>(Services.LOGGER);
+createTerminus(app, { healthChecks: { '/liveness': true }, onSignal: container.resolve('onSignal') });
+
+app.listen(port, () => {
+  logger.info(`app started on port ${port}`);
 });
